requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
batchgenerators==0.21
beautifulsoup4==4.11.1
bleach==5.0.1
blessings==1.7
cachetools==5.2.0
cffi==1.15.1
chainer==7.8.1
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
Cython==0.29.30
cython-npm==0.2.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dicom2nifti==2.4.3
dill==0.3.5.1
docker-pycreds==0.4.0
easydict==1.9
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
fcn==6.4.20
filelock==3.8.0
fonttools==4.34.4
frozenlist==1.3.1
fury==0.8.0
future==0.18.2
gdown==4.5.1
GeodisTK==0.1.7
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.47.0
h5py==3.7.0
humanize==4.4.0
idna==3.3
imageio==2.20.0
imcut==1.9.4
imgaug==0.4.0
importlib-metadata==3.10.1
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
MedPy==0.4.0
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nibabel==4.0.1
nnunet-customized==1.0
notebook==6.4.12
numpy==1.23.1
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openslide-python==1.2.0
p-tqdm==1.3.3
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.2.9
pathtools==0.1.2
pbd==0.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pox==0.3.1
ppft==1.7.6.5
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.3.0
pygco==0.0.16
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-gdcm==3.0.15
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.9.0
setproctitle==1.3.0
Shapely==1.8.4
shortuuid==1.0.9
SimpleITK==2.2.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
torch==1.10.1+cu111
torchaudio==0.10.1+rocm4.1
torchio==0.18.84
torchsummary==1.5.1
torchvision==0.11.2+cu111
tornado==6.2
TotalSegmentator==1.2
tqdm==4.64.0
traceback2==1.4.0
traitlets==5.3.0
typing_extensions==4.3.0
unittest2==1.1.0
urllib3==1.26.11
volumentations-3D==1.0.4
vtk==9.1.0
wandb==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.0
widgetsnbextension==3.6.1
wrapt==1.14.1
wslink==1.8.2
xvfbwrapper==0.2.9
yarl==1.8.1
zipp==3.8.1
